rename training_utils.py to feature_utils.py
