+++ plotting/stat_anal.py
@@ -3,7 +3,7 @@ from matplotlib import pyplot as plt
 
 import os
 
-import robustness_test.common_utils.training_utils
+import robustness_test.common_utils.feature_utils
 
 
 def correlation_stat_analysis(correlation, savefig=False, alternate_title=None, xlim=None, ylim=None):
@@ -65,7 +65,7 @@ if __name__ == "__main__":
     low_level_feature = "letters"
 
     # Load the data
-    path = robustness_test.common_utils.training_utils.get_prediction_path(language_model, feature, modality, subject, layer=layer)
+    path = robustness_test.common_utils.feature_utils.get_prediction_path(language_model, feature, modality, subject, layer=layer)
     print("loading", path)
     correlation = np.nan_to_num(np.load(path, allow_pickle=True))
     correlation_stat_analysis(correlation)
