+++ robustness_test/semantic_prediction.py
@@ -6,7 +6,7 @@ import numpy as np
 from himalaya.ridge import GroupRidgeCV
 from ridge_utils.util import make_delayed
 
-from robustness_test.common_utils.training_utils import load_downsampled_context_representations, load_subject_fmri, \
+from robustness_test.common_utils.feature_utils import load_downsampled_context_representations, load_subject_fmri, \
     get_prediction_path
 
 trim = 5
