+++ robustness_test/variance_partitioning.py
@@ -1,6 +1,6 @@
 import os
 
-from robustness_test.common_utils.training_utils import get_prediction_path
+from robustness_test.common_utils.feature_utils import get_prediction_path
 import numpy as np
 
 def ssc(data: np.array):
